% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx.R
\name{date_xx}
\alias{date_xx}
\alias{make_date_xx}
\alias{is_date_xx}
\title{A Superclass For All dint Objects}
\usage{
date_xx(x, subclass)

make_date_xx(y, q = NULL, m = NULL)

is_date_xx(x)
}
\arguments{
\item{x}{Any \R object}

\item{subclass}{subclass to assign}

\item{y, q, m}{Year, quarter, month. \code{q} and \code{m} are optional and at least one
of them must be \code{NULL}.}
}
\value{
a \code{date_xx} Object, except for \code{is_date_xx()} which returns \code{TRUE} or
\code{FALSE}

a \code{date_xx} Object for \code{date_xx()}, \code{make_date_xx}

\code{is_date_xx()} returns \code{TRUE} or \code{FALSE} depending on whether its
argument is of type \code{date_xx} or not.
}
\description{
Superclass for \code{date_xx} for \link{date_ym}, \link{date_yq} and \link{date_y}.

\code{make_date_xx} can be used to create such objects when it is not know if
month or quarter information is available.

\code{is_date_xx()} checks for \code{date_xx} objects.

\code{date_xx()} is an internally used constructor
that should only be used by developers aspiring to extend the dint package.
}
\examples{
make_date_xx(2017)
make_date_xx(2017, 4)
x <- make_date_xx(2017, m = 4)

is_date_xx(x)

}
