% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_ym.R, R/date_yq.R
\name{+.date_ym}
\alias{+.date_ym}
\alias{-.date_ym}
\alias{*.date_ym}
\alias{/.date_ym}
\alias{^.date_ym}
\alias{\%\%.date_ym}
\alias{\%/\%.date_ym}
\alias{seq.date_ym}
\alias{+.date_yq}
\alias{date_yq_arithmetic}
\title{Date Arithmetic Operations}
\usage{
\method{+}{date_ym}(x, y)

\method{-}{date_ym}(x, y)

\method{*}{date_ym}(x, y)

\method{/}{date_ym}(x, y)

\method{^}{date_ym}(x, y)

\method{\%\%}{date_ym}(x, y)

\method{\%/\%}{date_ym}(x, y)

\method{seq}{date_ym}(x, y, ...)

\method{+}{date_yq}(x, y)
}
\arguments{
\item{x}{a \code{\link{date_yq}} or \code{\link{date_ym}} object}

\item{y}{an integer}

\item{...}{currently ignored}
}
\description{
Currently only \code{+} and \code{-} are supported, all other basic arithmetic
operations are disabled for date_yq objects.
}
\seealso{
\link[base:Arithmetic]{base::Arithmetic}
}
