% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx.R, R/date_ym.R, R/date_yq.R
\name{date_xx_arithmetic_disabled}
\alias{date_xx_arithmetic_disabled}
\alias{*.date_ym}
\alias{/.date_ym}
\alias{^.date_ym}
\alias{\%\%.date_ym}
\alias{\%/\%.date_ym}
\alias{*.date_yq}
\alias{/.date_yq}
\alias{^.date_yq}
\alias{\%\%.date_yq}
\alias{\%/\%.date_yq}
\title{Disabled Date Arithmetic Operations}
\usage{
\method{*}{date_ym}(x, y)

\method{/}{date_ym}(x, y)

\method{^}{date_ym}(x, y)

\method{\%\%}{date_ym}(x, y)

\method{\%/\%}{date_ym}(x, y)

\method{*}{date_yq}(x, y)

\method{/}{date_yq}(x, y)

\method{^}{date_yq}(x, y)

\method{\%\%}{date_yq}(x, y)

\method{\%/\%}{date_yq}(x, y)
}
\arguments{
\item{x}{a \code{\link{date_yq}} or \code{\link{date_ym}} object}

\item{y}{an integer}
}
\description{
This page lists operators that are disabled for \code{date_yq} and \code{date_ym}
objects.
}
\seealso{
\link{date_xx_arithmetic}, \link[base:Arithmetic]{base::Arithmetic}
}
