% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx_scale.R
\name{date_xx_breaks_impl}
\alias{date_xx_breaks_impl}
\title{Pretty breaks for date_yq vectors}
\usage{
date_xx_breaks_impl(n = 6, padded = c(0L, 0L), period = 12L,
  as_dint = as_date_ym, get_subunit = get_month)
}
\arguments{
\item{n}{\code{NULL} or \code{integer} scalar}

\item{period}{\code{scalar} integer. number of units in a year (i.e 4 for
quarters, 12 for months, ...)}

\item{get_subunit}{}
}
\value{
a \code{function} that calculates a maximum of \code{n} breaks for a  \code{date_yq}
vector
}
\description{
Pretty breaks for date_yq vectors
}
