% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{date_xx_sequences}
\alias{date_xx_sequences}
\alias{seq.date_yw}
\alias{seq.date_yq}
\alias{seq.date_ym}
\title{date_xx Sequence Generation}
\usage{
\method{seq}{date_yw}(from, to, by = 1L, ...)

\method{seq}{date_yq}(from, to, by = 1L, ...)

\method{seq}{date_ym}(from, to, by = 1L, ...)
}
\arguments{
\item{from, to}{the starting and (maximal) end value of the sequence. Must be
of the same class (i.e. both must be a \code{date_yq}, \code{date_ym}, etc..)}

\item{by}{a positive \code{integer} scalar to increment the sequence with
(either in quarters, months or isoweeks, depending on the class of
\code{from}/\code{to})}

\item{...}{ignored}
}
\value{
an \code{integer} vector with the same \code{date_xx} subclass as \code{from}/\code{to}
}
\description{
date_xx Sequence Generation
}
