% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_y.R
\name{date_y}
\alias{date_y}
\alias{is_date_y}
\alias{as_date_y}
\title{A Simple S3-Class for Years}
\usage{
date_y(y)

is_date_y(x)

as_date_y(x)
}
\arguments{
\item{y}{year}

\item{x}{any R object}
}
\value{
\code{date_y} returns an object of type \code{date_y}

\code{is_date_y} returns \code{TRUE} or \code{FALSE} depending on whether its
argument is of type \code{date_y} or not.

\code{as_date_m} attempts to coerce its argument to \code{date_y} type
}
\description{
A simple data type for storing years. A \code{date_y} object is just an
integer with an additional class attribute.
}
\examples{
date_y(2013)

as_date_y(2016)

}
\concept{[date_xx] subclasses}
