% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx_scale.R
\name{date_ym_breaks}
\alias{date_ym_breaks}
\title{Pretty breaks for date_yq vectors}
\usage{
date_ym_breaks(n = 6)
}
\arguments{
\item{n}{\code{NULL} or \code{integer} scalar}

\item{padded}{\code{integer} vector of length 1 or two. Padding to remove
from each side of the input vector. This is useful for generating
breaks for ggplot objects where the limits are usually padded by
one month/quarter/isoweek for plotting.}

\item{period}{\code{scalar} integer. number of units in a year (i.e 4 for
quarters, 12 for months, ...)}

\item{get_subunit}{}
}
\value{
a \code{function} that calculates a maximum of \code{n} breaks for a  \code{date_yq}
vector
}
\description{
\code{date_xx_breaks_impl} does not return breaks, but a function that calculates
breaks. This is for compatbility with the breaks functions from scales such
as \code{\link[scales:pretty_breaks]{scales::pretty_breaks()}}, and for ease of use in ggplot.
}
