% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx_scale.R
\name{date_yq_breaks}
\alias{date_yq_breaks}
\title{Pretty breaks for date_yq vectors}
\usage{
date_yq_breaks(n = 6)
}
\arguments{
\item{n}{\code{NULL} or \code{integer} scalar}
}
\value{
a \code{function} that calculates a maximum of \code{n} breaks for a  \code{date_yq}
vector
}
\description{
Pretty breaks for date_yq vectors
}
