% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_of.R
\name{first_of_isoyear}
\alias{first_of_isoyear}
\alias{first_of_isoyear.default}
\alias{first_of_isoyear.date_yw}
\alias{first_of_isoyear.integer}
\alias{first_of_isoyear.numeric}
\alias{last_of_isoyear}
\alias{last_of_isoyear.default}
\alias{last_of_isoyear.date_yw}
\alias{last_of_isoyear.integer}
\alias{last_of_isoyear.numeric}
\title{#' Get First / Last Day of the First and Last Isoweek of a Year}
\usage{
first_of_isoyear(x)

\method{first_of_isoyear}{default}(x)

\method{first_of_isoyear}{date_yw}(x)

\method{first_of_isoyear}{integer}(x)

\method{first_of_isoyear}{numeric}(x)

last_of_isoyear(x)

\method{last_of_isoyear}{default}(x)

\method{last_of_isoyear}{date_yw}(x)

\method{last_of_isoyear}{integer}(x)

\method{last_of_isoyear}{numeric}(x)
}
\arguments{
\item{x}{anything that can be coerced to a \code{Date}}
}
\description{
#' Get First / Last Day of the First and Last Isoweek of a Year
}
