% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_day_of_month.R
\name{first_day_of_month}
\alias{first_day_of_month}
\alias{first_day_of_month.default}
\alias{last_day_of_month}
\alias{last_day_of_month.default}
\title{Get first / last day of a month}
\usage{
first_day_of_month(x)

\method{first_day_of_month}{default}(x)

last_day_of_month(x)

\method{last_day_of_month}{default}(x)
}
\arguments{
\item{x}{Anything that can be coerced to a date with \code{\link[base:as.Date]{base::as.Date()}}}
}
\value{
a \link{Date}
}
\description{
Get first / last day of a month
}
\examples{

first_day_of_month("2016-06-04")
last_day_of_month("2016-06-04")

}
