% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_of.R
\name{first_of_quarter}
\alias{first_of_quarter}
\alias{first_of_quarter.default}
\alias{last_of_quarter}
\alias{last_of_quarter.default}
\title{Get First / Last Day of a Quarter}
\usage{
first_of_quarter(x)

\method{first_of_quarter}{default}(x)

last_of_quarter(x)

\method{last_of_quarter}{default}(x)
}
\arguments{
\item{x}{Anything that can be coerced to a date with \code{\link[base:as.Date]{base::as.Date()}}}
}
\value{
a \link{Date}
}
\description{
Get First / Last Day of a Quarter
}
\examples{

first_of_quarter("2016-06-04")
last_of_quarter("2016-06-04")

}
