% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_day_of_year.R
\name{first_day_y}
\alias{first_day_y}
\alias{last_day_y}
\alias{first_day_of_year}
\alias{first_day_of_year.default}
\alias{last_day_of_year}
\alias{last_day_of_year.default}
\title{Get First or Last Day of a Year}
\usage{
first_day_y(x)

last_day_y(x)

first_day_of_year(x)

\method{first_day_of_year}{default}(x)

last_day_of_year(x)

\method{last_day_of_year}{default}(x)
}
\arguments{
\item{x}{Anything that can be coerced to a date with \code{\link[base:as.Date]{base::as.Date()}}}
}
\value{
a \link{Date}
}
\description{
\code{first_day_y()} is equivalent with \code{first_day_of_year()} and only included
for symmetry with \code{\link[=first_day_yq]{first_day_yq()}} and \code{\link[=first_day_ym]{first_day_ym()}}.
}
\examples{
first_day_y(2016)
first_day_y(2016)
first_day_of_year("2016-06-04")
last_day_of_year("2016-06-04")
}
\seealso{
\code{\link[=first_day_of_year]{first_day_of_year()}}

Other ym convenience functions: \code{\link{first_day_ym}},
  \code{\link{format_ym}}
}
\concept{ym convenience functions}
