% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{first_day_of_year}
\alias{first_day_of_month}
\alias{first_day_of_quarter}
\alias{last_day_of_year}
\alias{last_day_of_quarter}
\alias{last_day_of_month}
\title{Deprecated functions}
\usage{
first_day_of_year(...)

first_day_of_month(...)

first_day_of_quarter(...)

last_day_of_year(...)

last_day_of_quarter(...)

last_day_of_month(...)
}
\arguments{
\item{...}{See documentation of new function that replaces this one}
}
\description{
Deprecated functions
}
