% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dint-package.R
\docType{package}
\name{dint-package}
\alias{dint}
\alias{dint-package}
\title{dint: A Toolkit for Year-Quarter, Year-Month and Year-Isoweek
    Dates}
\description{
S3 classes and methods to create and work
    with year-quarter, year-month and year-isoweek vectors. Basic
    arithmetic operations (such as adding and subtracting) are supported,
    as well as formatting and converting to and from standard R date
    types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s-fleck/dint}
  \item Report bugs at \url{https://github.com/s-fleck/dint/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
