% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.date_xx}
\alias{format.date_xx}
\title{Format a date_xx}
\usage{
\method{format}{date_xx}(x, format, month_names = format(ISOdate(2000, 1:12,
  1), "\%B"), month_abb = format(ISOdate(2000, 1:12, 1), "\%b"), ...)
}
\arguments{
\item{x}{any \R object.}

\item{format}{A format that uses a subset of the same placeholders as
\code{\link[base:strptime]{base::strptime()}}:
\tabular{rl}{
\%Y \tab Year with century (the full year)\cr
\%y \tab Year without century (the last two digits of the year)\cr
\%m \tab Month as a decimal numbers (1-12)\cr
\%B \tab Full month name\cr
\%b \tab Abbreviated month name
}}

\item{month_names, month_abb}{a \code{character} vector of length 12: Names and
abbreviations for months that will be used for the
placeholders \code{"\%b"} and \code{"\%B"}. Defaults to the values for
the current locale for compatbilioty with \code{\link[base:strptime]{base::strptime()}}.}

\item{...}{ignored}
}
\description{
Format a date_xx
}
\examples{
x <- date_ym(2018, c(1L, 10L, 3L, 6L, 4L, 5L, 7L, 12L, 2L, 9L, 8L, 11L))
fm <- "\%Y-M\%m: \%B,\%b"

format.date_xx(
  x,
  format = fm,
  month_names = month.name,  # built-in R constant for English names
  month_abb = month.abb
)

}
