% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.date_yw}
\alias{format.date_yw}
\title{Format a date_yw Object}
\usage{
\method{format}{date_yw}(x, format = "\%Y-W\%W", ...)
}
\arguments{
\item{x}{a \link{date_yw} object}

\item{format}{A scalar character, valid values are: \code{"iso"}, \code{"short"}, and
\code{"shorter"}}

\item{...}{ignored}
}
\value{
A character vector
}
\description{
Format a date_yw Object
}
\examples{

x <- date_yw(2015, 12)

format(x, format = "iso")
# [1] "2015-M12"

format(x, format = "short")
# [1] "2015.12"

format(x, format = "shorter")
# [1] "15.12"

}
