% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_date_xx}
\alias{format_date_xx}
\alias{format_date_yq_iso}
\alias{format_date_yq_short}
\alias{format_date_yq_shorter}
\alias{format_date_ym_iso}
\alias{format_date_ym_short}
\alias{format_date_ym_shorter}
\title{Format shortcuts}
\usage{
format_date_yq_iso(x)

format_date_yq_short(x)

format_date_yq_shorter(x)

format_date_ym_iso(x)

format_date_ym_short(x)

format_date_ym_shorter(x)
}
\arguments{
\item{x}{anything that can be coerced to Date}
}
\description{
Format shortcuts
}
