% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment.R
\name{increment}
\alias{increment}
\alias{increment.date_yq}
\alias{increment.date_ym}
\alias{increment.date_yw}
\alias{increment.date_y}
\title{Increment date_xx objects}
\usage{
increment(x, inc = 1)

\method{increment}{date_yq}(x, inc)

\method{increment}{date_ym}(x, inc)

\method{increment}{date_yw}(x, inc)

\method{increment}{date_y}(x, inc)
}
\arguments{
\item{x}{object to increment}

\item{inc}{Value by which to increment (usually integer)}
}
\value{
An object of the same type as \code{x} increment by \code{inc}
}
\description{
Increment date_xx objects
}
