% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_date}
\alias{make_date}
\title{use the more efficient make_date function from lubridate if it is available}
\usage{
make_date(y, m, d)
}
\description{
use the more efficient make_date function from lubridate if it is available
}
