% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{Ops.date_xx}
\alias{Ops.date_xx}
\title{Comparison Operators for date_xx}
\usage{
\method{Ops}{date_xx}(e1, e2)
}
\arguments{
\item{e1, e2}{Objects with the same \code{date_xx} subclass (one of them can also
be integer)}
}
\value{
a \code{logical} scalar
}
\description{
Comparison Operators for date_xx
}
\examples{
date_yq(2015, 1) < date_yq(2015, 2)

# comparison with integers is ok
date_yq(2015, 1) < 20152

# but two different date_xx cannot be compared#'
try(date_yq(2015, 1) < date_ym(2015, 2))

}
