% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx.R
\name{print.date_xx}
\alias{print.date_xx}
\title{Print a date_xx Object}
\usage{
\method{print}{date_xx}(x, ...)
}
\arguments{
\item{x}{A \link{date_xx} object}

\item{...}{passed on to \code{\link[=format.date_yq]{format.date_yq()}} or \code{\link[=format.date_ym]{format.date_ym()}}}
}
\value{
\code{x} (invisibly)
}
\description{
Print a date_xx Object
}
