% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{rep.date_xx}
\alias{rep.date_xx}
\title{Replicate Elements of date_xx Vectors}
\usage{
\method{rep}{date_xx}(x, ...)
}
\arguments{
\item{x}{a \link{date_xx}}

\item{...}{passed on to \code{\link[base:rep]{base::rep()}}}
}
\value{
a vector of the same \code{date_xx} subclass as \code{x}
}
\description{
Replicate Elements of date_xx Vectors
}
