% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx_scale.R
\name{scale_date_xx}
\alias{scale_date_xx}
\alias{scale_x_date_yq}
\alias{scale_y_date_yq}
\alias{scale_x_date_ym}
\alias{scale_y_date_ym}
\alias{scale_x_date_yw}
\alias{scale_y_date_yw}
\title{ggplot2 scales for dint Objects}
\usage{
scale_x_date_yq(name = "Quarter", breaks = date_yq_breaks(),
  limits = NULL, position = "bottom")

scale_y_date_yq(name = "Quarter", breaks = date_yq_breaks(),
  limits = NULL, position = "left")

scale_x_date_ym(name = "Month", breaks = date_ym_breaks(),
  limits = NULL, position = "bottom")

scale_y_date_ym(name = "Month", breaks = date_ym_breaks(),
  limits = NULL, position = "left")

scale_x_date_yw(name = "Month", breaks = date_yw_breaks(),
  limits = NULL, position = "bottom")

scale_y_date_yw(name = "Month", breaks = date_yw_breaks(),
  limits = NULL, position = "left")
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the breaks specified by \code{date_breaks}
\item A \code{Date}/\code{POSIXct} vector giving positions of breaks
\item A function that takes the limits as input and returns breaks as output
}}

\item{limits}{A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum.}

\item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
}
\description{
The \code{scale_*_date_**} functions should provide nice defaults for plotting
the appropriate \link{date_xx} subclass, but come with a limited number of
configuration options. If you require more finetuning, you can convert
the appropriate vector with \code{as.Date()} and use \code{ggplot2::scale_x_date()}
}
\examples{
if (require("ggplot2", quietly = TRUE)){
  dd <- data.frame(date = seq(date_yq(2016, 1), date_yq(2018, 1)), V1 = 1:9)
  p <- ggplot(dd, aes(x = date, y = V1)) +
    geom_point()

  p  # automatically uses the proper scale
  p + scale_x_date_yq("quarters with default spacing")
  p + scale_x_date_yq(breaks = date_yq_breaks(3))
}

}
\seealso{
\link{date_xx_breaks}
}
