% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{Summary.date_xx}
\alias{Summary.date_xx}
\title{Maxima and Minima for date_xx}
\usage{
\method{Summary}{date_xx}(..., na.rm)
}
\arguments{
\item{...}{\code{date_xx} vectors with the same subclass}

\item{na.rm}{logical: should missing values be removed?}
}
\value{
for \code{min()} and \code{max()} a scalar of the same \code{date_xx} subclass as it's
input, for range a vector of length 2
}
\description{
Maxima and Minima for date_xx
}
\examples{
min(date_yq(2014, 1), date_yq(2014, 2))

# raises an error
try(min(date_yq(2014, 1), date_ym(2014, 2)))
}
