% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{\%y+\%}
\alias{\%y+\%}
\alias{\%y-\%}
\alias{\%y+\%.date_y}
\alias{\%y-\%.date_y}
\alias{\%y+\%.date_yq}
\alias{\%y-\%.date_yq}
\alias{\%y+\%.date_ym}
\alias{\%y-\%.date_ym}
\alias{\%y+\%.date_yw}
\alias{\%y-\%.date_yw}
\title{Add/Subtract Year}
\usage{
x \%y+\% y

x \%y-\% y

\method{\%y+\%}{date_y}(x, y)

\method{\%y-\%}{date_y}(x, y)

\method{\%y+\%}{date_yq}(x, y)

\method{\%y-\%}{date_yq}(x, y)

\method{\%y+\%}{date_ym}(x, y)

\method{\%y-\%}{date_ym}(x, y)

\method{\%y+\%}{date_yw}(x, y)

\method{\%y-\%}{date_yw}(x, y)
}
\arguments{
\item{x}{a \link{date_xx} vector}

\item{y}{an \code{integer} vector of years}
}
\description{
Add/Subtract Year
}
\examples{
date_yq(2017, 1) \%y+\% 1
date_yq(2017, 1) \%y-\% 1
date_ym(2017, 1) \%y+\% 1
date_ym(2017, 1) \%y-\% 1

}
