% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{yq}
\alias{yq}
\alias{qy}
\title{Parse dates with year and quarter components}
\usage{
yq(x)

qy(x)
}
\arguments{
\item{x}{a \code{character} vector}
}
\value{
a \code{date_yq} vector
}
\description{
Parse dates with year and quarter components
}
\examples{
yq("2018 1")
qy("1st Quarter 2019")

}
