% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo-yearqtr.R
\name{as_yearqtr}
\alias{as_yearqtr}
\alias{as_yearqtr.date_yq}
\alias{as_yearqtr.yearqtr}
\alias{as_yearmon}
\alias{as_yearmon.date_ym}
\alias{as_yearmon.yearmon}
\alias{as_yearweek}
\alias{as_yearweek.date_yw}
\alias{as_yearweek.yearweek}
\title{Coerce to zoo yeartqr objects}
\usage{
as_yearqtr(x)

\method{as_yearqtr}{date_yq}(x)

\method{as_yearqtr}{yearqtr}(x)

as_yearmon(x)

\method{as_yearmon}{date_ym}(x)

\method{as_yearmon}{yearmon}(x)

as_yearweek(x)

\method{as_yearweek}{date_yw}(x)

\method{as_yearweek}{yearweek}(x)
}
\arguments{
\item{x}{any \R object}
}
\value{
a \link[zoo:yearqtr]{zoo::yearqtr} vector
}
\description{
\code{as_yearqtr()} is included for interoperatility with \code{\link[zoo:yearqtr]{zoo::yearqtr()}},
an alternative year-quarter format that is based on a decimal representation
as opposed to dint's integer representation of year-quarters.
}
\examples{

x <- date_yq(2016, 2)
as_yearqtr(x)

}
