% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx.R, R/date_ym.R, R/date_yq.R,
%   R/date_yw.R
\name{date_xx_arithmetic}
\alias{date_xx_arithmetic}
\alias{+.date_ym}
\alias{-.date_ym}
\alias{seq.date_ym}
\alias{+.date_yq}
\alias{-.date_yq}
\alias{seq.date_yq}
\alias{+.date_yw}
\alias{-.date_yw}
\alias{seq.date_yw}
\title{Date Arithmetic Operations}
\usage{
\method{+}{date_ym}(x, y)

\method{-}{date_ym}(x, y)

\method{seq}{date_ym}(x, y, ...)

\method{+}{date_yq}(x, y)

\method{-}{date_yq}(x, y)

\method{seq}{date_yq}(x, y, ...)

\method{+}{date_yw}(x, y)

\method{-}{date_yw}(x, y)

\method{seq}{date_yw}(x, y, ...)
}
\arguments{
\item{x}{a \code{\link{date_yq}} or \code{\link{date_ym}} object}

\item{y}{an integer}

\item{...}{currently ignored}
}
\description{
The arithmetic operations \code{+}, \code{-} as well as sequence generation with
\code{seq()} are all supported for \code{date_yq} and \code{date_ym} objects. Other binary
arithmetic operators are disabled (see \link{date_xx_arithmetic_disabled}).
}
\examples{
q <- date_yq(2018, 1)
q + 5
q - 1
seq(q, q + 5)


m <- date_ym(2018, 12)
m + 1
m - 13
seq(m - 1, m + 1)
}
\seealso{
\link[base:Arithmetic]{base::Arithmetic}
}
