% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{date_xx_arithmetic_disabled}
\alias{date_xx_arithmetic_disabled}
\alias{*.date_xx}
\alias{/.date_xx}
\alias{^.date_xx}
\alias{\%\%.date_xx}
\alias{\%/\%.date_xx}
\title{Disabled Date Arithmetic Operations}
\usage{
\method{*}{date_xx}(x, y)

\method{/}{date_xx}(x, y)

\method{^}{date_xx}(x, y)

\method{\%\%}{date_xx}(x, y)

\method{\%/\%}{date_xx}(x, y)
}
\arguments{
\item{x}{a \code{\link{date_yq}} or \code{\link{date_ym}} object}

\item{y}{an integer}
}
\description{
This page lists operators that are disabled for \code{date_yq} and \code{date_ym}
objects.
}
\seealso{
\link{date_xx_arithmetic}, \link[base:Arithmetic]{base::Arithmetic}
}
