% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_of.R
\name{first_of_year}
\alias{first_of_year}
\alias{first_of_year.integer}
\alias{first_of_year.default}
\alias{first_of_year.numeric}
\alias{last_of_year}
\alias{last_of_year.integer}
\alias{last_of_year.default}
\alias{last_of_year.numeric}
\title{Get First / Last Day of a Year}
\usage{
first_of_year(x)

\method{first_of_year}{integer}(x)

\method{first_of_year}{default}(x)

\method{first_of_year}{numeric}(x)

last_of_year(x)

\method{last_of_year}{integer}(x)

\method{last_of_year}{default}(x)

\method{last_of_year}{numeric}(x)
}
\arguments{
\item{x}{Anything that can be coerced to a date with \code{\link[base:as.Date]{base::as.Date()}}}
}
\value{
a \link{Date}
}
\description{
Get First / Last Day of a Year
}
\examples{
first_of_year("2016-06-04")
last_of_year("2016-06-04")
}
