% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_date_xx}
\alias{format_date_xx}
\alias{format_yq_iso}
\alias{format_yq_short}
\alias{format_yq_shorter}
\alias{format_ym_iso}
\alias{format_ym_short}
\alias{format_ym_shorter}
\alias{format_yw_iso}
\alias{format_yw_short}
\alias{format_yw_shorter}
\title{Format Shortcuts}
\usage{
format_yq_iso(x)

format_yq_short(x)

format_yq_shorter(x)

format_ym_iso(x)

format_ym_short(x)

format_ym_shorter(x)

format_yw_iso(x)

format_yw_short(x)

format_yw_shorter(x)
}
\arguments{
\item{x}{anything that can be coerced to \code{Date}}
}
\value{
a \code{character} vector
}
\description{
Functions that apply formatting presets to anything that can be
coerced to a \code{date_xx}. This is notably handy as it can be used as a
labelling function for \strong{ggplot2} (see \code{vignette("dint")})
}
