% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_yw.R
\name{format_yw}
\alias{format_yw}
\title{Directly Create Formatted Year-Week Strings}
\usage{
format_yw(x, w = NULL, format = "iso")
}
\arguments{
\item{x, m}{Two integer (vectors). \code{m} is optional and the interpretation of
\code{x} will depend on whether \code{m} is supplied or not:
\itemize{
\item if only \code{x} is supplied, \code{x} will be passed to \code{\link[=as_date_yw]{as_date_yw()}}
(e.g. \code{x = 201604} means April 2016)
\item if \code{x} and \code{m} are supplied, \code{x} is interpreted as year and \code{m} as
week.
}}

\item{format}{A scalar character, valid values are: \code{"iso"}, \code{"short"}, and
\code{"shorter"}}
}
\value{
A character vector
}
\description{
Directly Create Formatted Year-Week Strings
}
\examples{

format_yw(2015, 5)
format_yw(201505, format = "short")
format_yw(201505, format = "shorter")

}
\seealso{
\code{\link[=format.date_yw]{format.date_yw()}}
}
\concept{yw convenience functions}
