% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_of.R
\name{first_of_isoweek}
\alias{first_of_isoweek}
\alias{first_of_isoweek.default}
\alias{last_of_isoweek}
\alias{last_of_isoweek.default}
\alias{first_of_isoyear}
\alias{first_of_isoyear.date_yw}
\alias{first_of_isoyear.integer}
\alias{first_of_isoyear.numeric}
\alias{last_of_isoyear}
\alias{last_of_isoyear.date_y}
\title{Get first / last day of a year}
\usage{
first_of_isoweek(x)

\method{first_of_isoweek}{default}(x)

last_of_isoweek(x)

\method{last_of_isoweek}{default}(x)

first_of_isoyear(x)

\method{first_of_isoyear}{date_yw}(x)

\method{first_of_isoyear}{integer}(x)

\method{first_of_isoyear}{numeric}(x)

last_of_isoyear(x)

\method{last_of_isoyear}{date_y}(x)
}
\arguments{
\item{x}{Anything that can be coerced to a date with \code{\link[base:as.Date]{base::as.Date()}}}
}
\value{
a \link{Date}
}
\description{
Get first / last day of a year
}
\examples{
first_of_isoweek("2016-06-04")
last_of_isoweek("2016-06-04")
}
