% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{Summary.date_xx}
\alias{Summary.date_xx}
\title{Maxima and Minima for date_xx}
\usage{
\method{Summary}{date_xx}(..., na.rm)
}
\arguments{
\item{...}{date_xx vectors. It is up to the user to ensure that they are
of the same subclass.}

\item{na.rm}{}
}
\value{
for \code{min()} and \code{max()} a scalar of the same \code{date_xx} subclass as it's
input, for range a vector of length 2
}
\description{
Maxima and Minima for date_xx
}
