% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_xx.R, R/date_ym.R, R/date_yq.R
\name{\%y+\%}
\alias{\%y+\%}
\alias{\%y-\%}
\alias{\%y+\%.date_ym}
\alias{\%y-\%.date_ym}
\alias{\%y+\%.date_yq}
\alias{\%y-\%.date_yq}
\title{Add/Subtract Year}
\usage{
x \%y+\% y

x \%y-\% y

\method{\%y+\%}{date_ym}(x, y)

\method{\%y-\%}{date_ym}(x, y)

\method{\%y+\%}{date_yq}(x, y)

\method{\%y-\%}{date_yq}(x, y)
}
\description{
Add/Subtract Year
}
\examples{
date_yq(2017, 1) \%y+\% 1
date_yq(2017, 1) \%y-\% 1
date_ym(2017, 1) \%y+\% 1
date_ym(2017, 1) \%y-\% 1

}
